refactor(api): make export schema from api
