+++ packages/api/src/schema/index.ts
@@ -1,0 +1,2 @@
+export * from './connection-params';
+export * from './schema';
